requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
argcomplete==1.10.3
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.8.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
compressed-rtf==1.0.6
dataclasses-json==0.5.9
decorator==5.1.1
docx2txt==0.8
ebcdic==1.1.1
extract-msg==0.28.7
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
idna==3.4
IMAPClient==2.1.0
importlib-metadata==6.7.0
Jinja2==3.1.2
jsonschema==4.17.3
langchain==0.0.222
langchainplus-sdk==0.0.20
lxml==4.9.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.0
olefile==0.46
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.3
pdfminer.six==20191110
Pillow==9.5.0
protobuf==4.23.3
pyarrow==12.0.1
pycryptodome==3.18.0
pydantic==1.10.10
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pypdf==3.12.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-pptx==0.6.21
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
reportlab==4.0.4
requests==2.31.0
rich==13.4.2
six==1.12.0
smmap==5.0.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SpeechRecognition==3.8.1
SQLAlchemy==2.0.17
streamlit==1.24.0
streamlit-chat==0.1.1
tenacity==8.2.2
textract==1.6.5
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.3
validators==0.20.0
watchdog==3.0.0
xlrd==1.2.0
XlsxWriter==3.1.2
yarl==1.9.2
zipp==3.15.0
